Refactor after changes to core. Can now suport 2 way comms with the backround tasks.

+++ src/webpack/plugin.initTransformer.js
@@ -10,13 +10,15 @@ class PluginInitTransformer {
 
   apply (compiler) {
     compiler.hooks.done.tap('done-compiling', async () => {
-      const manData = fse.readFileSync(this.options.chunkManifest)
-      const data = fse.readFileSync(path.join(this.options.src, 'js', this.options.fileName))
-      const varData = `const BEXChunkData=${manData.toString()};`
-      let newValue = data.toString().indexOf('BEXChunkData=') > -1
-        ? data.toString().replace(/const[ .*]BEXChunkData.*};/g, varData)
-        : varData + data.toString()
-      fse.writeFileSync(path.join(this.options.src, 'js', this.options.fileName), newValue, 'utf-8')
+      if (fse.existsSync(this.options.chunkManifest)) {
+        const manData = fse.readFileSync(this.options.chunkManifest)
+        const data = fse.readFileSync(path.join(this.options.src, 'js', this.options.fileName))
+        const varData = `const BEXChunkData=${manData.toString()};`
+        let newValue = data.toString().indexOf('BEXChunkData=') > -1
+          ? data.toString().replace(/const[ .*]BEXChunkData.*};/g, varData)
+          : varData + data.toString()
+        fse.writeFileSync(path.join(this.options.src, 'js', this.options.fileName), newValue, 'utf-8')
+      }
     })
   }
 }
